applied stackedNav in homepage
