+++ src/routes/pages/grades/+page.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
 
 
-// import AristomateNav from '$components/shared/AristomateNav.svelte';
 import StackedNav from "$components/shared/StackedNav"
 import Grades from '$components/grades/gradesPage.svelte';
 </script>
