+++ src/routes/quickLinks/quickLinks.svelte
@@ -83,8 +83,8 @@
 	});
 </script>
 
-<SubPageHeader title="Χρήσιμοι σύνδεσμοι" />
-<ion-content>
+<SubPageHeader title="Χρήσιμοι σύνδεσμοι" stackedNav />
+<ion-content class="ion-padding">
 	{#each links as { linktitle, linkaddress, imagelink }}
 		<LinkCard {linktitle} {linkaddress} {imagelink} />
 	{/each}
